+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [9.2.0] (IN PROGRESS)
 
 * Agreements accordion - Revise Are you sure you want to unassign agreement? message. (UIEH-1420)
+* Agreements accordion - Add a tooltip for the `New` and `Add` buttons. (UIEH-1424)
 
 ## [9.1.1] (https://github.com/folio-org/ui-eholdings/tree/v9.1.1) (2024-03-24)